requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.16.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.6
install==1.3.5
lxml==5.1.0
numpy==1.26.3
pandas==2.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
rsa==4.9
six==1.16.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
